Detect fee issues when sent amount is higher than received amount
When the sent amount is higher than the received amount, we now assume a
fee was paid and adjust the transaction accordingly, unless a fee was
already set in a different currency. Warnings are currently printed to
console, but should of course be part of the UI.

Also rendered the auto-detected "Buy" and "Sell" transaction types in
italic to make them stand out a little, given that these might indicate
missing wallets if they were really just transfers.

Closes #5