Merge consecutive trades to reduce UI clutter
When selling some amount of crypto on an exchange, it is often executed
as a series of orders. For tax declaration purposes this should be
irrelevant, so now we merge such consecutive trades.

Whether transactions are allowed to merge in this way should probably be
an option, either per portfolio or per source.