Improved handling of sources that have a long import process
Now, the sync is no longer on startup, but can be triggered manually,
with the results getting stored in the sources file. Ideally it would
also be done async, but that's a change for another time.

Also improved the Bitcoin Address importing to support wallets with
multiple addresses.

Finally, when a transaction is sending to the same wallet (it can
happen...), it is now imported as a plain "Fee" transaction.