Added support for reading Binance CSV exports
This adds support for importing some of the information that can be
exported from Binance. Unfortunately, many services provided by Binance
have their own export files, so still excludes many areas of Binance
which I personally did not use.

Now we can import:

* Wallet Transaction History CSV
* Spot Trade History CSV
* BNB Convert History, once manualy copy/pasted from the website into
  the right CSV format.

The string to Amount parsing was improved so that it can handle no space
between the quantity and the currency, as well as commas in the
quantity.