Added initial support for tracking NFTs
For now, an optional "token_id" field was added to Amount, with the
contract name/address for the token stored in the "currency" field. For
tracking holdings, the token id and constract are concatenated with a
":" in between, since when we sell the token we need to calculate the
gain using that specific token.

Also implemented importing such amounts from the Trezor Suite Ethereum
CSV export. But unfortunately, the CSV file is very inconsistent with
RECV and SENT in the Type column, making it hard or impossible to know
whether a token is actually sent or received.