Added importing of Stellar accounts
Using the Horizon API through the stellar-horizon crate, though pointing
at the salessandri/stellar-horizon-rs fork because the one at aurora-rs
appears to be unmaintained.

The import of Stellar account transactions is based on the "Tax Data
Export" feature of StellarExpert[1,2], extended to also take into
account claimed balances.

Also changed the Bitcoin imports to async, though the UI is still
blocking on it for now.

[1] https://stellar.expert/tax-export/public
[2] https://github.com/stellar-expert/stellar-expert-explorer/blob/master/ui/business-logic/tax-info-exporter.js