Don't block the UI while syncing
Use `slint::spawn_local` instead of `futures::executor::block_on`. Since
`slint::spawn_local` polls the future on the main thread, this is the
easiest way to avoid the async code from blocking the UI, since it does
not require introducing any Arc / Mutex and allows us to keep the Rc
members in App.

The UI still needs to provide some feedback when a sync is in progress.