Fixed 100% CPU usage and freezing system while synchronizing
For some reason the async API calls into Horizon and possibly other APIs
could start taking 100% CPU and eventually freeze the app or even the
entire system when left running for a while.

This happened already when using the futures::executor::block_on method
to block on the async call, and was still happening when using
slint::spawn_local to poll the future on the event loop. Only now that
we run the future on another thread with tokio::task::spawn does this
issue seem to be gone.