Added "Export All" action
It asks for a directory and will export the reports for each year there
in CSV format, in addition to the "all_time" reports.

Also extended the Report Summary CSV export with the actual summary that
is also displayed in the UI, in addition to the existing per-currency
inventory summary.