Added links from Capital Gain report to transactions
Now you can click the timestamp of either the buy or the sell part of
the capital gain event, to quickly jump to the relevant transaction.

You can also select transactions now and navigate them with up, down,
page-up and page-down, but this doesn't really have a purpose yet, apart
from it being used to highlight a transaction when clicking from the
Capital Gain report.

Increased the space between table columns from 2px to 6px to improve
readability.