Fixed ETH deposits to Bittrex not getting recognized
Bittrex export appears to show a different transaction hash for ETH
deposits, since they show the hash of a second "Sweep" transaction
rather than the transaction to the deposit address.

Since transactions with different hashes are currently never combined
into a transfer, we don't set the hash on ETH deposits as a workaround.