Added support for Swap transactions
The Swap transaction is used to swap one currency for another, while
retaining the original cost base and holding period. It can be used when
a symbol gets renamed or for stake / unstake transactions.

For now it is only automatically applied to 1INCH / st1INCH tokens, when
imported from an Ethereum address.