Improved accuracy of Poloniex CSV import
Actually this change introduces some rounding to zero on the calculated
total amount traded and the imported fee.

This makes the import more accurate because it appears Poloniex is doing
the same kind of rounding, even though the fees are written out more
precise in some of their exported files.