Added SwapIn and SwapOut transaction types to CTC
This change is just adding them for completeness, but they are not
actually supported yet since in Raccoin a swap is a single entry whereas
the SwapIn and SwapOut need to come in pairs.

Maybe CryptoTaxCalculator will make it possible to use the base and
quote amounts with a single CSV entry instead.