Replaced slint::invoke_from_event_loop with slint::spawn_local
By using slint::spawn_local, we avoid needing to pass a reference to the
UI or the App through an async future. As such we can switch back from
using Arc<Mutex<App>> to Rc<RefCell<App>>.

This change reverts most of 679fcfa48f57f168329bb395e57b64aaa9659bdc.
The part that remains is the usage of tokio::task::spawn for the async
part.

While it isn't recommended to use `#[tokio::main]`, I've kept it for now
and added tokio::task::block_in_place usage as shown in the Slint
documentation, because the rfd crate is configured to use Tokio for its
xdg-portal feature and I'm not sure if it would work without
`#[tokio::main]` at the moment.