Add per-wallet cost basis tracking to FIFO accounting
In "universal" cost basis tracking, the FIFO algorithm could entirely
ignore holding transfers. Now it also supports "per-wallet" cost basis
tracking, where it uses `wallet_index` of matched send/receive
transactions to determine which lot is affected by the transfer.

The `wallet_index` can be looked up by the global transaction index
stored in `matching_tx`. However, since `FIFO::process` already received
a mutable reference to a slice of the transactions, Rust does not allow
it to also get a non-mutable reference to the entire vector. For now,
this problem is solved by making a temporary copy of this information in
the `tx_meta` vector.