Adjusted demo to show a difference in cost basis tracking
Previously the demo was set up such that the oldest BTC was being sold
regardless of how the cost basis was calculated. Now the transactions
are ordered such that per-wallet cost basis tracking moves a profitable
trade to short-term, which in this case means a lot more taxes would be
due (but one can also make examples where per-wallet tracking is
preferable).