Switched to Skia renderer on Linux (#22)
It supports both hardware and software rendering at decent speed and
renders text much nicer. It does introduce a dependency on fontconfig.

We keep using FemtoVG on macOS for now because it seems to perform
somewhat better.

We also keep using FemtoVG on Windows for now, because in testing I got
only a white window that closed after a few seconds.

Also updated the dependencies, though staying with Slint 1.13 for now
because of a regression in the FemtoVG renderer (slint-ui/slint#10078).