Improved error handling and display (#46)
When there was an error while handling a Swap (like insufficient balance
or missing cost base), make sure the supposedly available amount is
still added to the holdings. This avoids the failed handling of the Swap
causing additional balance errors.

For transfers, only errors associated with the Send transaction were
visible in the UI. However, FIFO could also assign an error to the
Receive transaction. Now either error is displayed.

Fixed some transactions in the demo portfolio which were actually
impossible due to insufficient balance.