Try to handle loss of funds on arrival by setting a "receive fee"
Sometimes a blockchain has a higher precision than an exchange and small
amounts can get lost on the receiving end. This can happen for ETH
deposits to Bitstamp for example, which seem to get truncated to 8
decimal places even though ETH tokens can have up to 18 decimal places.

To avoid small balance errors, we now set the lost amount as a "receive
fee". The display of the Transfer transaction in the UI has been
adjusted such that it will sum up the send and receive fees.

Also, Transfers now display in the list at the timestamp of their
receiving end. This matches the time at which the FIFO process handles
the transfer, so any balance errors related to the processing order are
now less confusing.