Don't treat airdrops as income (#53)
This should probably be an option eventually, but in Germany airdrops
can generally be treated as a zero-cost buy (exception is if it's more
like a bounty).

Treating airdrops as a zero-cost buy is often preferable because it can
be hard to determine their value at the time of the drop and it can be
very volatile.